refactor :: CompaniesViewModel 추상화
