+++ feature/company/src/main/java/team/retum/company/viewmodel/CompaniesViewModel.kt
@@ -5,12 +5,18 @@ import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.FlowPreview
+import kotlinx.coroutines.flow.debounce
+import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import team.retum.common.base.BaseViewModel
 import team.retum.usecase.entity.CompaniesEntity
 import team.retum.usecase.usecase.company.FetchCompaniesUseCase
 import team.retum.usecase.usecase.company.FetchCompanyCountUseCase
 import javax.inject.Inject
+
+private const val SEARCH_DEBOUNCE_MILLIS = 1000L
 
 @HiltViewModel
 internal class CompaniesViewModel @Inject constructor(
@@ -22,20 +28,47 @@ internal class CompaniesViewModel @Inject constructor(
     internal val companies: List<CompaniesEntity.CompanyEntity> = _companies
 
     init {
-        fetchTotalCompanyPageCount()
-        fetchCompanies()
+        debounceName()
+    }
+
+    @OptIn(FlowPreview::class)
+    private fun debounceName() {
+        viewModelScope.launch {
+            state.map { it.name }.distinctUntilChanged().debounce(SEARCH_DEBOUNCE_MILLIS).collect {
+                setState { state.value.copy(showCompaniesEmptyContent = it?.isNotEmpty() ?: false) }
+                if (!it.isNullOrBlank()) {
+                    fetchTotalCompanyPageCount()
+                    fetchCompanies()
+                }
+            }
+        }
+    }
+
+    internal fun setName(name: String) {
+        val initialState = CompaniesState.getDefaultState()
+        _companies.clear()
+        setState {
+            state.value.copy(
+                name = name,
+                page = initialState.page,
+                totalPage = initialState.totalPage,
+            )
+        }
     }
 
     internal fun fetchCompanies() {
         viewModelScope.launch(Dispatchers.IO) {
-            fetchCompaniesUseCase(
-                page = state.value.page,
-                name = null,
-            ).onSuccess {
-                addPage()
-                _companies.addAll(it.companies)
-            }.onFailure {
-                postSideEffect(CompaniesSideEffect.FetchCompaniesError)
+            with(state.value) {
+                fetchCompaniesUseCase(
+                    page = page,
+                    name = name,
+                ).onSuccess {
+                    setState { copy(showCompaniesEmptyContent = it.companies.isEmpty()) }
+                    addPage()
+                    _companies.addAll(it.companies)
+                }.onFailure {
+                    postSideEffect(CompaniesSideEffect.FetchCompaniesError)
+                }
             }
         }
     }
@@ -48,9 +81,9 @@ internal class CompaniesViewModel @Inject constructor(
         setState { copy(page = page + 1) }
     }
 
-    private fun fetchTotalCompanyPageCount() {
+    internal fun fetchTotalCompanyPageCount() {
         viewModelScope.launch(Dispatchers.IO) {
-            fetchCompanyCountUseCase(name = null).onSuccess {
+            fetchCompanyCountUseCase(name = state.value.name).onSuccess {
                 setState { state.value.copy(totalPage = it.totalPageCount) }
             }
         }
@@ -60,11 +93,15 @@ internal class CompaniesViewModel @Inject constructor(
 internal data class CompaniesState(
     val totalPage: Long,
     val page: Int,
+    val name: String?,
+    val showCompaniesEmptyContent: Boolean,
 ) {
     companion object {
         fun getDefaultState() = CompaniesState(
             totalPage = 0,
             page = 1,
+            name = null,
+            showCompaniesEmptyContent = false,
         )
     }
 }
